Version 0.22 Переменован целый пакет tasks в botprocess. Требуется отладка

+++ src/main/java/ru/staffbot/webserver/servlets/ControlServlet.java
@@ -2,19 +2,16 @@ package ru.staffbot.webserver.servlets;
 
 import ru.staffbot.database.Database;
 import ru.staffbot.database.configs.Configs;
-import ru.staffbot.database.settings.Settings;
-import ru.staffbot.utils.levers.BooleanLever;
+import ru.staffbot.utils.botprocess.BotProcess;
+import ru.staffbot.utils.botprocess.BotProcessStatus;
 import ru.staffbot.utils.levers.Lever;
 import ru.staffbot.utils.levers.Levers;
-import ru.staffbot.utils.tasks.TaskStatus;
-import ru.staffbot.utils.tasks.Tasks;
 import ru.staffbot.utils.values.ValueType;
 import ru.staffbot.webserver.AccountService;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -32,16 +29,15 @@ public class ControlServlet extends MainServlet {
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         Map<String, Object> pageVariables = new HashMap();
 
-        for (TaskStatus status : TaskStatus.values()) {
+        for (BotProcessStatus status : BotProcessStatus.values()) {
             String statusName = "control_" + status.name().toLowerCase();
-            if (status == TaskStatus.PAUSE)
-                pageVariables.put(statusName, TaskStatus.START == Tasks.getStatus() ? "" : "disabled");
+            if (status == BotProcessStatus.PAUSE)
+                pageVariables.put(statusName, BotProcessStatus.START == BotProcess.getStatus() ? "" : "disabled");
             else
-                pageVariables.put(statusName, status == Tasks.getStatus() ? "disabled" : "");
+                pageVariables.put(statusName, status == BotProcess.getStatus() ? "disabled" : "");
         }
 
-//        pageVariables.put("control_status", Tasks.getStatus());
-        pageVariables.put("control_start_time", Long.toString(Tasks.getStartTime()));
+        pageVariables.put("control_start_time", Long.toString(BotProcess.getStartTime()));
         pageVariables.put("control_display", Database.connected() ? "inline-table" : "none");
         pageVariables.put("page_bg_color", page_bg_color);
         pageVariables.put("control_leverlist", getLeverList());
@@ -55,20 +51,22 @@ public class ControlServlet extends MainServlet {
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         if (request.getParameter("control_apply") == null) {
             // Обработка управляющих кнопок (пуск, пауза, старт)
-            for (TaskStatus status : TaskStatus.values())
+            for (BotProcessStatus status : BotProcessStatus.values())
                 if (request.getParameter("control_" + status.name().toLowerCase()) != null)
-                    Tasks.setStatus(status);
+                    BotProcess.setStatus(status);
 
             // Обработка кнопок для работы с конфигурацией (сохранить, загрузить, удалить)
             String configName = request.getParameter("control_configname");
             if (configName != null)
-                try {
-                    Configs config = new Configs(configName);
-                    if (request.getParameter("control_save") != null) config.save();
-                    if (request.getParameter("control_load") != null) config.load();
-                    if (request.getParameter("control_delete") != null) config.delete();
-                } catch (Exception e) {}
-
+                if (!configName.equals(""))
+                    try {
+                        Configs config = new Configs(configName);
+                        if (request.getParameter("control_save") != null) config.save();
+                        if (request.getParameter("control_load") != null) config.load();
+                        if (request.getParameter("control_delete") != null) config.delete();
+                    } catch (Exception e) {} finally {
+                        BotProcess.reSchedule();
+                    }
         } else {
             for (Lever lever : Levers.list) {
                 if (lever.getValueType() == ValueType.EMPTY) continue;
@@ -76,8 +74,9 @@ public class ControlServlet extends MainServlet {
                 String leverValue = request.getParameter(leverName);
                 lever.setValueFromString(leverValue);
             }
-            Tasks.reSchedule();
+            BotProcess.reSchedule();
         }
+
         doGet(request, response);
     }
 
@@ -101,8 +100,8 @@ public class ControlServlet extends MainServlet {
             for (String configName : Configs.getList()) {
                 context += "<option value=\"" + configName + "\">";
             }
-        } catch (Exception e) {
-            //e.printStackTrace();
+        } catch (Exception exception) {
+            context = "";
         }
         return context;
     }
